requirements.txt
streamlit
pandas
chromadb
openai
google-genai
openpyxl
python-dotenv
sentence-transformers
scikit-learn
